+++ src/main/ts/ephox/alloy/api/behaviour/Tooltipping.ts
@@ -2,13 +2,15 @@ import { TooltippingBehaviour } from '../../behaviour/tooltipping/TooltippingTyp
 import TooltippingSchema from '../../behaviour/tooltipping/TooltippingSchema';
 import * as TooltippingState from '../../behaviour/tooltipping/TooltippingState';
 import * as ActiveTooltipping from '../../behaviour/tooltipping/ActiveTooltipping';
+import * as TooltippingApis from '../../behaviour/tooltipping/TooltippingApis';
 import * as Behaviour from './Behaviour';
 
 const Tooltipping = Behaviour.create({
   fields: TooltippingSchema,
   name: 'tooltipping',
   active: ActiveTooltipping,
-  state: TooltippingState
+  state: TooltippingState,
+  apis: TooltippingApis
 }) as TooltippingBehaviour;
 
 export {