+++ src/main/ts/ephox/alloy/behaviour/sandboxing/SandboxApis.ts
@@ -32,6 +32,14 @@ const open = (sandbox: AlloyComponent, sConfig: SandboxingConfig, sState: Sandbo
   return newState;
 };
 
+// TODO AP-191 write a test for openWhileCloaked
+const openWhileCloaked = (sandbox: AlloyComponent, sConfig: SandboxingConfig, sState: SandboxingState, data, transaction: () => void) => {
+  cloak(sandbox, sConfig, sState);
+  open(sandbox, sConfig, sState, data);
+  transaction();
+  decloak(sandbox, sConfig, sState);
+};
+
 const close = (sandbox: AlloyComponent, sConfig: SandboxingConfig, sState: SandboxingState) => {
   sState.get().each((data) => {
     Attachment.detachChildren(sandbox);
@@ -89,6 +97,7 @@ export {
   cloak,
   decloak,
   open,
+  openWhileCloaked,
   close,
   isOpen,
   isPartOf,