+++ src/main/ts/ephox/alloy/behaviour/tooltipping/TooltippingTypes.ts
@@ -1,7 +1,7 @@
 import { BehaviourConfigSpec, BehaviourConfigDetail, AlloyBehaviour } from '../../api/behaviour/Behaviour';
 import { AlloyComponent } from '../../api/component/ComponentApi';
 import { Result, Option } from '@ephox/katamari';
-import { RawDomSchema } from '../../api/component/SpecTypes';
+import { RawDomSchema, AlloySpec } from '../../api/component/SpecTypes';
 import { BehaviourState } from '../../behaviour/common/BehaviourState';
 
 export interface TooltippingBehaviour extends AlloyBehaviour<TooltippingConfigSpec, TooltippingConfig> {
@@ -12,6 +12,7 @@ export interface TooltippingConfig extends BehaviourConfigDetail {
   // Make consistent with other lazy sinks.
   lazySink: () => (comp: AlloyComponent) => Result<AlloyComponent, any>;
   tooltipDom: () => RawDomSchema;
+  tooltipComponents: () => AlloySpec[];
   exclusive: () => boolean;
   delay: () => number;
 }
@@ -20,6 +21,7 @@ export interface TooltippingConfigSpec extends BehaviourConfigSpec {
   // Make consistent with other lazy sinks.
   lazySink: (comp: AlloyComponent) => Result<AlloyComponent, any>;
   tooltipDom: RawDomSchema;
+  tooltipComponents?: AlloySpec[];
   exclusive?: boolean;
   delay?: number;
 }