+++ src/test/ts/browser/ui/inline/InlineViewTest.ts
@@ -1,4 +1,4 @@
-import { Assertions, GeneralSteps, Logger, Mouse, Step, UiFinder, Waiter } from '@ephox/agar';
+import { Assertions, GeneralSteps, Logger, Mouse, Step, UiFinder, Waiter, Chain } from '@ephox/agar';
 import { UnitTest } from '@ephox/bedrock';
 import { Arr, Future, Option, Result } from '@ephox/katamari';
 import * as GuiFactory from 'ephox/alloy/api/component/GuiFactory';
@@ -11,6 +11,7 @@ import * as TestDropdownMenu from 'ephox/alloy/test/dropdown/TestDropdownMenu';
 import * as GuiSetup from 'ephox/alloy/test/GuiSetup';
 import * as Sinks from 'ephox/alloy/test/Sinks';
 import * as TestBroadcasts from 'ephox/alloy/test/TestBroadcasts';
+import { Html, Css } from '@ephox/sugar';
 
 UnitTest.asynctest('InlineViewTest', (success, failure) => {
 
@@ -86,6 +87,15 @@ UnitTest.asynctest('InlineViewTest', (success, failure) => {
 
     return [
       UiFinder.sNotExists(gui.element(), '.test-inline'),
+
+      Logger.t(
+        'Check that getContent is none for an inline menu that has not shown anything',
+        Step.sync(() => {
+          const contents = InlineView.getContent(inline);
+          Assertions.assertEq('Should be none', true, contents.isNone());
+        })
+      ),
+
       Step.sync(() => {
         InlineView.showAt(inline, {
           anchor: 'selection',
@@ -98,10 +108,57 @@ UnitTest.asynctest('InlineViewTest', (success, failure) => {
       }),
       sCheckOpen('After show'),
 
+      Logger.t(
+        'Check that getContent is some now that the inline menu has shown something',
+        Step.sync(() => {
+          const contents = InlineView.getContent(inline);
+          Assertions.assertEq('Checking HTML of inline contents', 'Inner HTML', Html.get(contents.getOrDie(
+            'Could not find contents'
+          ).element()));
+        })
+      ),
+
       Step.sync(() => {
         InlineView.hide(inline);
       }),
 
+      Logger.t(
+        'Check that getContent is none not that inline view has been hidden again',
+        Step.sync(() => {
+          const contents = InlineView.getContent(inline);
+          Assertions.assertEq('Should be none', true, contents.isNone());
+        })
+      ),
+
+      sCheckClosed('After hide'),
+
+      Step.sync(() => {
+        InlineView.showAt(inline, {
+          anchor: 'makeshift',
+          x: 50,
+          y: 50
+        }, Container.sketch({
+          dom: {
+            innerHtml: 'Inner HTML'
+          }
+        }));
+      }),
+      sCheckOpen('After show'),
+
+      Logger.t(
+        'Check that inline view has a top and left',
+        Chain.asStep(gui.element(), [
+          UiFinder.cFindIn('.test-inline'),
+          Chain.op((value) => {
+            Assertions.assertEq('Check view CSS top is 50px', '50px', Css.getRaw(value, 'top').getOr('no top found'));
+            Assertions.assertEq('Check view CSS left is 50px', '50px', Css.getRaw(value, 'left').getOr('no left found'));
+          })
+        ])
+      ),
+
+      Step.sync(() => {
+        InlineView.hide(inline);
+      }),
       sCheckClosed('After hide'),
 
       Logger.t(