+++ changelog.md
@@ -6,6 +6,50 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.24.0] - 2018-09-04
+### Added
+ - Added InlineView.showMenuAt() to special-case positioning for inline menus
+ - Added Sandboxing.openWhileCloaked() convenience method
+
+## [3.23.0] - 2018-08-31
+### Added
+ - Added eventOrder for Dropdowns
+ - Added extra debugging information
+
+## [3.22.0] - 2018-08-29
+### Added
+ - Added dragging behaviour flag for repositionTarget (defaults to true)
+ - Added dragging behaviour handler (onDrag)
+
+## [3.21.0] - 2018-08-29
+### Added
+ - onChoose event to Sliders
+
+## [3.20.0] - 2018-08-28
+### Added
+ - Replacing.replaceAt and Replacing.replaceBy
+
+## [3.19.0] - 2018-08-23
+### Added
+ - Tooltipping API access to hideAllExclusive, and tooltipComponents in config
+ - DomFactory.simple and DomFactory.dom for quick generation of basic AlloySpec objects
+ - InlineView API: getContent
+ - Readable state for Flatgrid Keying types
+ - Support for matrix-style menus
+ - Consistent definitions for itemBehaviours and widgetBehaviours
+ - IgnoreFocus capability for item widgets
+ - Exposing onChangeTab and onDismissTab through TabSectionTypes
+ - Chain methods for TestStore
+
+## [3.18.0] - 2018-08-20
+### Added
+ - selectClasses and selectAttributes to HtmlSelect sketcher
+
+## [3.17.0] - 2018-08-10
+### Added
+ - Configuration for InlineView: fireDismissalEventInstead
+ - SystemEvents.dismissRequested()
+
 ## [3.16.0] - 2018-08-08
 ### Added
 - Reflecting behaviour